Fixes a few small oof with AI names (#2625)
* Fixes a few small oof with AI names

I'm stoopid
h


e


f

* kek

* Update code/datums/jobs/job/job.dm


* h

+++ code/datums/jobs/job/shipside.dm
@@ -706,6 +706,9 @@ As a Synthetic you answer to the acting captain. Special circumstances may chang
 	if(!length(GLOB.ai_spawn))
 		CRASH("attempted to spawn an AI with no landmark set")
 
+	if(!H.mind) //Could be a dummy.
+		return
+
 	var/mob/living/silicon/ai/AI = new(pick(GLOB.ai_spawn))
 	H.mind.transfer_to(AI, TRUE)
 	qdel(H)
@@ -713,7 +716,7 @@ As a Synthetic you answer to the acting captain. Special circumstances may chang
 
 /datum/job/ai/assign(mob/living/carbon/human/H, visualsOnly, announce, latejoin, datum/outfit/outfit_override, client/preference_source)
 	if(preference_source?.prefs)
-		H.fully_replace_character_name(H.name, preference_source.prefs.synthetic_name)
+		H.fully_replace_character_name(H.name, preference_source.prefs.ai_name)
 
 	return ..()
